[FW:H:V2.47] 1. Modified Setup_Audio() and Init_CODEC() to support standard I2S format on UIF 2. Fixed bug of miss to read PIO_ISR in some case in Check_GPIO_Intrrupt() 3. Fixed bug in error check in parse_to_host_command()

+++ Noah/im501_comm.h
@@ -89,7 +89,6 @@ typedef struct {
 
 unsigned char MCU_Load_Vec( unsigned char firsttime );
 
-
 unsigned char im501_read_reg_i2c( unsigned char reg_addr, unsigned char *pdata );
 unsigned char im501_read_reg_spi( unsigned char reg_addr, unsigned char *pdata );
 
@@ -110,6 +109,6 @@ unsigned char test_send_cmd_to_im501( void );
 
 unsigned char Read_iM501_Voice_Buffer( unsigned char gpio_irq, unsigned int timeout_ms, unsigned char pkt_sn );
 
-unsigned char Write_CMD_To_iM501( unsigned char cmd_index, unsigned int para );
+unsigned char Write_CMD_To_iM501( unsigned char cmd_index, unsigned short para );
 
 #endif
